Adding filestore download utility, minor updates to handle file sizes larger than 2GB correctly
Former-commit-id: 1f80a3765b1a1809bdcc2268f3ef1f9c5b32ecc7

+++ src/java/edu/ucsd/library/dams/file/impl/SwiftClient.java
@@ -353,7 +353,7 @@ public class SwiftClient
 		BufferedOutputStream out = new BufferedOutputStream(
 			new FileOutputStream(fn)
 		);
-		int bytesRead = -1;
+		long bytesRead = -1L;
 		try
 		{
 			bytesRead = FileStoreUtil.copy( in, out );
